# documentation/openapi.yaml
openapi: 3.0.1
info:
  title: Musicman API
  description: Описание всех эндпоинтов сервиса "Musicman"
  contact:
    name: Анастасия Сашина
    email: sashina@cs.vsu.ru
  version: 1.0.6
servers:
  - url: http://localhost:8080/api/v1
    description: Generated server url
security:
  - Bearer: []
paths:
...
  "/auth/profile":
    put:
      tags:
        - Auth Controller
      summary: Обновление информации о пользователе
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateProfileRequest"
        required: true
      responses:
        '200':
          description: Информация обновлена
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Нет доступа
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '409':
          description: Пользователь c указанным ником уже существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
...
  "/auth/verify/{code}":
    post:
      tags:
        - Auth Controller
      summary: Верификация пользователя
      operationId: verifyUser
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Пользователь верифицирован
          content:
            application/json: {}
        '404':
          description: Ссылка невалидна
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '405':
          description: Истекло время действия ссылки
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
      security: []
  "/auth/resend-code":
    post:
      tags:
        - Auth Controller
      summary: Повторная отправка кода
      operationId: resendCode
      responses:
        '200':
          description: Код повторно отправлен
          content:
            application/json: {}
        '403':
          description: Нет доступа
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '405':
          description: Пользователь уже верифицирован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
  "/auth/register":
    post:
      tags:
        - Auth Controller
      summary: Регистрация нового пользователя
      operationId: signUpNewUserUsingForm
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserSignUpRequest"
        required: true
      responses:
        '201':
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TokensDTO"
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '409':
          description: Пользователь с указанной почтой или указанным ником уже существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
      security: []
  "/auth/refresh-token":
    post:
      tags:
        - Auth Controller
      summary: Обновление access токена
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        '200':
          description: Возвращение обновленных токенов
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TokensDTO"
        '400':
          description: Токен отсутствует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '401':
          description: Токен не валиден
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Токен не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
      security: []
  "/auth/login":
    post:
      tags:
        - Auth Controller
      summary: Вход пользователя
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LoginRequest"
        required: true
      responses:
        '200':
          description: Пользователь успешно вошел в систему
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TokensDTO"
        '400':
          description: Неправильный пароль или почта
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
      security: []
  "/auth/credentials":
    post:
      tags:
        - Auth Controller
      summary: Получение информации о пользователе
      operationId: getCredentials
      responses:
        '200':
          description: Получение кредов
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CredentialsDTO"
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Нет доступа
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
...
components:
  schemas:
    ExceptionDTO:
      type: object
      properties:
        message:
          type: string
...
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          maxLength: 100
          minLength: 0
          type: string
        lastName:
          maxLength: 100
          minLength: 0
          type: string
        nickname:
          maxLength: 100
          minLength: 0
          type: string
...
    TokensDTO:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    UserSignUpRequest:
      required:
        - email
        - firstName
        - lastName
        - nickname
        - password
        - repeatPassword
      type: object
      properties:
        firstName:
          maxLength: 100
          minLength: 0
          type: string
        lastName:
          maxLength: 100
          minLength: 0
          type: string
        email:
          type: string
        nickname:
          maxLength: 100
          minLength: 0
          type: string
        password:
          maxLength: 20
          minLength: 8
          type: string
        repeatPassword:
          maxLength: 20
          minLength: 8
          type: string
    RefreshTokenRequest:
      required:
        - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
    LoginRequest:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    CredentialsDTO:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        nickname:
          type: string
        isVerified:
          type: boolean
        roles:
          uniqueItems: true
          type: array
          items:
            "$ref": "#/components/schemas/RoleDTO"
    RoleDTO:
      type: object
      properties:
        roleId:
          type: integer
          format: int64
        name:
          type: string
        authority:
          type: string
...
  securitySchemes:
    Authorization:
      type: http
      scheme: Bearer
      bearerFormat: JWT